feat: wallet endpoint
